requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Babel
Flask-Caching
Flask-Migrate==4.0.5
GeoAlchemy2
psycopg2-binary
requests
scikit-learn
tensorflow
geopandas
pandas
numpy
python-dotenv
gunicorn
SQLAlchemy
psycopg2
flask_alembic